+++ ngbatis-demo/src/main/java/ye/weicheng/ngbatis/demo/pojo/vertex/Player.java
@@ -1,27 +1,40 @@
-package ye.weicheng.ngbatis.demo.pojo.Vertex;
+package ye.weicheng.ngbatis.demo.pojo.vertex;
 
 // Copyright (c) 2022 All project authors. All rights reserved.
 //
 // This source code is licensed under Apache 2.0 License.
 
 import org.nebula.contrib.ngbatis.annotations.base.GraphId;
-import org.nebula.contrib.ngbatis.enums.IdType;
 import org.nebula.contrib.ngbatis.annotations.base.Tag;
 import org.nebula.contrib.ngbatis.base.GraphBaseVertex;
+import org.nebula.contrib.ngbatis.enums.IdType;
+
 
 /**
- * Team点实体示例
+ * Player点实体示例
  * @author xYLiuuuuuu
  * @since 2024/9/11
  */
 
-@Tag(name = "team")
-public class Team extends GraphBaseVertex {
+@Tag(name = "player")
+public class Player extends GraphBaseVertex {
 
   @GraphId(type = IdType.STRING)
+  private String id;
+
   private String name;
 
-  public String getName() {
+  private Integer age;
+
+  String getId() {
+    return id;
+  }
+
+  public void setId(String id) {
+    this.id = id;
+  }
+
+  String getName() {
     return name;
   }
 
@@ -29,10 +42,22 @@ public class Team extends GraphBaseVertex {
     this.name = name;
   }
 
+  Integer getAge() {
+    return age;
+  }
+
+  public void setAge(Integer age) {
+    this.age = age;
+  }
+
   @Override
   public String toString() {
-    return "Team{" +
+    return "Player{" +
             "name='" + name + '\'' +
+            ", age=" + age +
             '}';
   }
+
+
+
 }
